refactor(google): move autoscaling policy interfaces and layout components to autoscalingPolicy directory
Currently, the only Typescript/React components of the autoscaling policy modal are specific to the scale-in control feature. In the next commit, we will add support for predictive autoscaling, so in order to share these interfaces and layout components, lift them to the top level.
