+++ app/scripts/modules/google/src/autoscalingPolicy/IGceAutoscalingPolicy.ts
@@ -1,0 +1,11 @@
+export interface IGceAutoscalingPolicy {
+  scaleInControl: IGceScaleInControl;
+}
+
+export interface IGceScaleInControl {
+  maxScaledInReplicas?: {
+    fixed?: number;
+    percent?: number;
+  };
+  timeWindowSec?: number;
+}
