+++ app/scripts/modules/google/src/autoscalingPolicy/gceAutoscalingFieldLayout.less
@@ -1,5 +1,5 @@
 // todo(mneterval): remove when GCE Autoscaling Controls are entirely converted to React & Formik
-.gce-scale-in-controls {
+.gce-autoscaling-layout {
   .checkbox {
     margin-left: 5px;
     margin-top: 7px;
