+++ app/scripts/modules/google/src/autoscalingPolicy/index.ts
@@ -1,0 +1,2 @@
+export * from './IGceAutoscalingPolicy';
+export * from './GceAutoScalingFieldLayout';
